Reuse existing pvc fixture in minio namespace
